# api/openapi.yaml
openapi: 3.0.3
info:
  title: Entrevistes API
  version: 1.0.0
  description: API mínima para alumnes, entrevistes, cursos, import/export.
servers:
  - url: https://gei.adeptify.es/api
    description: Producción (on-prem)
  - url: https://REEMPLAZA-CON-TU-URL-CLoudRun.run.app
    description: Cloud Run
paths:
...
  /alumnes/{id}:
    get:
      summary: Fitxa completa de l'alumne
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlumneFull'
        '404': { $ref: '#/components/responses/NotFound' }
        '403': { $ref: '#/components/responses/Forbidden' }
    put:
      summary: Actualiza datos no sensibles (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlumneUpdate'
      responses:
        '200': { $ref: '#/components/responses/UpdatedWithId' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  responses:
    BadRequest:
      description: Dades requerides incompletes
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: 'Dades requerides incompletes' }
    Forbidden:
      description: Permís denegat
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: 'Permís denegat' }
    NotFound:
      description: No trobat
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: 'No trobat' }
...
    UpdatedWithId:
      description: Updated
      content:
        application/json:
          schema:
            type: object
            properties:
              id: { type: string }
              status: { type: string, example: 'updated' }
  schemas:
...
    AlumneUpdate:
      type: object
      properties:
        grup: { type: string }
        estat: { type: string, enum: [alta, baixa, migrat] }
    DadesPersonals:
      type: object
      properties:
        sexe: { type: string }
        dataNaixement: { type: string, format: date }
    Tutor:
      type: object
      properties:
        nom: { type: string }
        telefon: { type: string }
        email: { type: string }
    AlumneFull:
      type: object
      properties:
        id: { type: string }
        nom: { type: string }
        grup: { type: string }
        anyCurs: { type: string }
        estat: { type: string }
        dadesPersonals: { $ref: '#/components/schemas/DadesPersonals' }
        dadesFamiliars:
          type: object
          properties:
            tutors:
              type: array
              items: { $ref: '#/components/schemas/Tutor' }
...